Re-layout on Changing Display Name.
Fixes https://github.com/lxde/lxqt/issues/1061.

When the sorting is not by name, If the display name of a desktop entry or shortcut file on the desktop changes without change in its file name, a re-layout should be done; otherwise, the item might be out of grid until the next desktop refresh.